+++ models/controls/rapport_controles_geo_as_point.sql
@@ -3,7 +3,8 @@
     tags=['grace_rapport'],
     post_hook = ["ALTER TABLE {{ this }} ADD PRIMARY KEY (id);"],
     indexes = [
-      {'columns': ['geom'], 'type': 'gist'}
+      {'columns': ['geom'], 'type': 'gist'},
+      {'columns': ['criticite'], 'type': 'btree'}
     ]
   )
 }}
@@ -35,6 +36,7 @@ select
   attribut,
   id_entite,
   detail_erreur,
+  criticite,
   case
     when geom is null or st_isempty(geom) then null
     when geometrytype(geom) in ('POINT', 'MULTIPOINT')